# Fix/perf issue