--- a/src/acc/libsmm_acc/libcusmm/libcusmm.h
+++ b/src/acc/libsmm_acc/libcusmm/libcusmm.h
@@ -16,9 +16,12 @@
 #include <cuda.h>
 #include <cuda_runtime.h>
 #include <nvrtc.h>
-#include <omp.h>
 #include <unordered_map>
 #include <vector>
+
+#if defined _OPENMP
+#include <omp.h>
+#endif
 
 // Macros for CUDA error handling
 // Wrap calls to CUDA NVRTC API
@@ -48,7 +51,9 @@ struct kernel_launcher {
 };
 
 static std::unordered_map<Triplet, kernel_launcher> kernel_handles;
+#if defined _OPENMP
 static std::unordered_map<Triplet, omp_lock_t> kernel_locks;
+#endif
 
 int libcusmm_process_d(int *param_stack, int stack_size,
     CUstream stream, int m, int n, int k,