@shoshijak 2019-02-05
Please note this PR is **not ready** to be merged.

A race condition appears when two threads try to JIT two different `(m, n, k)`s simultaneously (observed, for example, when running `dbcsr_unittest3` with #ranks >=2, #threads >= 12). I'm still working on this. 

This race condition is introduced by this PR, there is no race condition in the current `dbcsr/develop`.